+++ system/library/nuvei/ocmod_includes/catalog_header.php
@@ -6,12 +6,12 @@ try {
     $this->load->model('setting/setting');
     
     // add Nuvei Checkout SDK
-    if('prod' == $this->config->get(NUVEI_SETTINGS_PREFIX . 'sdk_version')) {
+//    if('prod' == $this->config->get(NUVEI_SETTINGS_PREFIX . 'sdk_version')) {
         $this->document->addScript(NUVEI_SDK_URL_PROD);
-    }
-    else {
-        $this->document->addScript(NUVEI_SDK_URL_INT);
-    }
+//    }
+//    else {
+//        $this->document->addScript(NUVEI_SDK_URL_INT);
+//    }
     
     // add Nuvei common modify script
     $this->document->addScript('catalog/view/javascript/nuvei_common_js_mod.js');